Remove erroneous extra currency code.
To my understanding - the currency lines in liquidity (_get_dst_liquidity_aml_dict_vals) are just wrong and should not be here at all. I understand the need to refactor to get operating unit in, but base Odoo does not have this code about currency.
Now, this fails if the currency and the company currency do not match the journal / target currency.
E.g. A company in Australian Currency transfers 100 Euro from a Euro account to a USD account. When writing to the USD side, self.currency_id will be Euro and self.company_id.currency_id will be AUD, but the transaction will be expecting the USD equivalent.
Base Odoo is currenly handling it all correctly in '_get_shared_move_line_vals', but this code breaks it.